Add support for pre connection hooks
It may be desirable for the user to be able to apply certain settings
to a QSqlDatabase before use in QDjango, however, since connections are
cloned on a per-thread basis there is currently no way to do this.
Introducing pre-connection hooks allows the user to complete any neccessary
configuration of the database connection before ORM use.